+++ src/components/Button.php
@@ -4,6 +4,7 @@
 namespace vippsas\login\components;
 
 use Craft;
+use craft\helpers\StringHelper;
 
 class Button
 {
@@ -88,6 +89,12 @@ class Button
      */
     private $type = self::TYPE_LOGIN;
 
+    /**
+     * Optional return url
+     * @var string
+     */
+    private $return_url;
+
     // Public Methods
     // =========================================================================
 
@@ -122,7 +129,9 @@ class Button
         else $a = ' ' . $a;
         if ($img == null) $img = '';
         else $img = ' ' . $img;
-        return "<a href=\"{$this->href}\"{$a}><img src=\"/vipps/asset/button/{$filename}\"{$img}></a>";
+        if($this->return_url) $href = $this->href . '?r=' . $this->return_url;
+        else $href = $this->href;
+        return "<a href=\"{$href}\"{$a}><img src=\"/vipps/asset/button/{$filename}\"{$img}></a>";
     }
 
     /**
@@ -241,6 +250,16 @@ class Button
     public function pill()
     {
         $this->shape = self::SHAPE_PILL;
+        return $this;
+    }
+
+    /**
+     * @param $url
+     * @return $this
+     */
+    public function returnUrl($url)
+    {
+        $this->return_url = StringHelper::base64UrlEncode($url);
         return $this;
     }