+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Support for "Automatic Return" using PKCE
+- `returnUrl()` function for button to redirect the user to a specific URL on return
 
 ## 1.0.0 - 2020-11-14